- Create a basic logic file - Fix more ESlint errors
